# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Online Store API, probably",
    "description": "API for user registration, login, profile management, and administrative actions."
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/categories/post": {
    "post":{
      "tags": ["categories"],
      "summary": "post category",
      "description": "posts category with provied category id, name, description",
      "parameters": [
        {
          "in": "body",
          "name": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "name": {"type": "string"},
              "description": {"type": "string"},
              "parent_category_id": {"type": "integer"}
            }
          }
         }
      ],
      "security": [
              {
                "BearerAuth": []
              }
            ],
            "responses": [
              {
                "201": {"description": "category has been created successfully"}, 
                "400": {"description": "missing required arguments"},
                "409": {"description": "no duplicate category names allowed"}
              }
            ]
    }
  },
  "/categories/get": {
    "get": {
      "tags": ["categories"],
      "summary": "get all categories", 
      "description": "fetches all existing categories",
      "security": [
        {
          "BearerAuth": []
        }
      ],
      "responses": [
        {
          "200": {"description": "List of users returned"},
          "401": {"description": "access denied"}
        }
      ]
    }
  },
  "/categories/get/{id}":{
    "get": {
      "tags": ["categories"],
      "summary": "get category by id",
      "description": "fetches category at the given id",
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "type": "integer",
          "required": true
        }
      ],
      "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": [
          {
            "200": {"description": "returned category at the given id"},
            "404": {"description": "category at the given id not found"}
          }
        ]
    }
  },
  "/categories/put/{id}":{
    "put": {
      "tags": ["categories"],
      "summary": "post category",
      "description": "updates at the given id, missing fields get reset",
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "type": "integer",
          "required": true
        },
        {
          "in": "body",
          "name": "body",
          "schema": {
            "type": "object",
            "properties": {
              "name": {"type": "string"},
              "description": {"type": "string"},
              "parent_category_id": {"type": "integer"}
            }
          }
        }
      ],
      "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": [
          {
            "201": {"description": "updated category at the given id"},
            "404": {"description": "category at the given id not found"}
          }
        ]
    }
  },
  "/categories/patch/{id}":{
    "patch": {
      "tags": ["categories"],
      "summary": "patch category",
      "description": "updates at the given id preserving existing data",
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "type": "integer",
          "required": true
        },
        {
          "in": "body",
          "name": "body",
          "schema": {
            "type": "object",
            "properties": {
              "name": {"type": "string"},
              "description": {"type": "string"},
              "parent_category_id": {"type": "integer"}
            }
          }
        }
      ],
      "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": [
          {
            "201": {"description": "updated category at the given id"},
            "404": {"description": "category at the given id not found"}
          }
        ]
    }
  },
  "/categories/delete/{id}":{
    "delete": {
      "tags": ["categories"],
      "summary": "delete category",
      "description": "deletes user at the given id",
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "type": "integer",
          "required": true
        }
      ],
      "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": [
          {
            "201": {"description": "deleted category at the given id. serial sequence reset to min available"},
            "404": {"description": "category at the given id not found"}
          }
        ]
    }
  },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
    }
  }
}